Add support for models with albedo tint
In Source, it is possible to change the color of a prop entity by simply changing `rendercolor`. In Call of Duty, however, having the same model in different colors requires having duplicates of the model with the color applied on its material(s). Support for this feature is added for Black Ops 3 only as older CoD titles are unlikely to handle that many duplicate models and materials.