Rewrite the way assets are exported and named to avoid duplicates that will overwrite each other
In Source, it is possible to store your assets in different folders. This makes it possible for multiple assets to have the same name. If we export them using their name only, duplicates are bound to happen and that will cause many problems that are not possible to fix. The best way I could think of to fix that issue is to export them with names with their paths hashed to make sure they are exported with unique names.