Fix what I broke while trying to fix the unique name issue
For some reason, I broke a lot of things while implementing what I thought would be the best idea to avoid duplicate names in models and materials. I'm undoing that with in this commit, but I'll go back to it after receiving some feedback while the beta-testing phase is still going on.