Create ModelConverter.py
The model converter I wrote for the previous version had a lot of problems. It was a separate executable, so it had to be called for each model, it was 70 mb in size and most importantly, it didn't convert vertex normals properly. This file uses the model loader in SourceMDL and solves all those problems while working as a part of Corvid.