Change the way assets are named/renamed when they are exported
In Source, you can store your assets in subfolders inside models and materials folders, which allows you to have different assets with the same name in different folders. However, this was a problem for Corvid since you need to have all the materials, models and images in their respective folders with no subfolders in Call of Duty, and if you imported a map with those assets, some assets were likely to overwrite others because of the said problem. After the changes that came with this commit, Corvid will export all the assets with the hashed version of their folders in their names to avoid name clashes.