Slice patches that are 17x17 in size into 4 9x9 patches.
CoD 4 does not support terrain patches bigger than 16x16 and Radiant crashes with no error messages when it tries to open a map with them. With this update, Corvid now checks for the size of the terrain patches and what game the map is being converted for, and it them into 9x9 terrain patches if it's bigger than 16x16 and the map is being converted for Cod 4.